+++ Changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Client and Server Side Event to atomic/atom complete design. Completed on 2025-11-27
+- Add load components package.json content into cengine class at compmgr engine. Completed on 2025-12-07
 
 ### Changed